requirements.txt
openai>=1.3.0
anthropic>=0.8.0
google-generativeai>=0.3.0
cohere>=4.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio
rich>=13.0.0
click>=8.0.0
pydantic>=2.0.0
tiktoken>=0.5.0 